requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
requests
plotly
streamlit
streamlit-option-menu
scrapy
pydantic
python-dateutil
email-validator
